bus/vmbus: add hyper-v virtual bus support
This patch adds support for an additional bus type Virtual Machine BUS
(VMBUS)is used on Microsoft Hyper-V in Windows 10, Windows Server 2016
and Azure.  Most of this code was extracted from FreeBSD and some of
this is from earlier code donated by Brocade.

Only Linux is supported at present, but the code is split
to allow future FreeBSD and Windows support.

This version supports multiple channels per device and requires
requires this bus support as well revised versio of uio_hv_generic
which is has been submitted to upstream Linux.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>