mk: fix build shared library when --sysroot is defined
If doing cross compile it is common to use --sysroot in
EXTRA_CFLAGS to use a different root for includes and c runtime.
When a shared library is linked using CC this flag needs to
be passed in order for libraries to be found.

It also necessary if doing link time optimization.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>