net/netvsc: always handle events
Since VF state changes occur on the primary channel, and there
is no gurantee that the primary channel (queue 0) will get polled;
need to always poll the primary channel on every Tx and Rx.

Add some overhead, but quite small since if ring is empty, no
lock is acquired; and if another core is processing the queue
the lock can be skipped.

This means the concept of tx_free_thresh is irrelevant
for this driver, and that checking and local variable
can be removed.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>